+++ crates/module-system/module-implementations/sov-test-modules/src/lib.rs
@@ -3,3 +3,4 @@ pub mod cache_module;
 pub mod gas;
 pub mod hooks_count;
 pub mod pinned_cache;
+pub mod sequencing_data;