+++ crates/module-system/sov-modules-api/src/module/spec.rs
@@ -8,7 +8,7 @@ use sov_rollup_interface::crypto::{CredentialId, Signature};
 use sov_rollup_interface::da::DaSpec;
 use sov_rollup_interface::optimistic::Attestation;
 use sov_rollup_interface::zk::{CryptoSpec, StateTransitionPublicData, Zkvm};
-use sov_rollup_interface::BasicAddress;
+use sov_rollup_interface::{BasicAddress, Bytes};
 use sov_state::{Storage, StorageProof};
 
 use crate::gas::Gas;
@@ -160,6 +160,8 @@ pub struct Context<S: Spec> {
     sequencer: S::Address,
     /// The DA layer address of the sequencer who included the transaction.
     sequencer_da_address: <S::Da as DaSpec>::Address,
+    /// Sequencing data provided by the sequencer
+    sequencing_data: Option<Bytes>,
     /// The rollup address that pays the gas fees for the transaction.
     gas_refund_recipient: S::Address,
 }
@@ -180,6 +182,11 @@ impl<S: Spec> Context<S> {
         &self.sequencer_da_address
     }
 
+    /// Returns the sequencing data
+    pub fn sequencing_data(&self) -> &Option<Bytes> {
+        &self.sequencing_data
+    }
+
     /// Returns the rollup address which will receive any gas refund from the transaction.
     pub fn gas_refund_recipient(&self) -> &S::Address {
         &self.gas_refund_recipient
@@ -196,6 +203,7 @@ impl<S: Spec> Context<S> {
         sender_credentials: Credentials,
         sequencer: S::Address,
         sequencer_da_address: <S::Da as DaSpec>::Address,
+        sequencing_data: Option<Bytes>,
     ) -> Self {
         Self::with_payer(
             sender,
@@ -203,6 +211,7 @@ impl<S: Spec> Context<S> {
             sequencer,
             sequencer_da_address,
             sender,
+            sequencing_data,
         )
     }
 
@@ -213,6 +222,7 @@ impl<S: Spec> Context<S> {
         sequencer: S::Address,
         sequencer_da_address: <S::Da as DaSpec>::Address,
         payer: S::Address,
+        sequencing_data: Option<Bytes>,
     ) -> Self {
         Self {
             sender_credentials,
@@ -220,6 +230,7 @@ impl<S: Spec> Context<S> {
             sequencer,
             sequencer_da_address,
             gas_refund_recipient: payer,
+            sequencing_data,
         }
     }
 
@@ -266,6 +277,7 @@ mod arbitrary {
                 Default::default(),
                 sequencer,
                 sequencer_da_address,
+                None,
             ))
         }
     }